Make safe_request decorate generic to reuse
The generic version of the safe_request decorator can then be
specialized to use a particular error via functools.partial, thus
avoiding code duplication
